Fix errors in production build
The fix required rollup to be updated and the only way to update rollup
is to also update vite. Other packages needed to also be updated as a
result.

The issue is because the library `react-audio-player` is bundled with
webpack which exposes an object. The older versions of vite/rollup are
not able to handle this correctly. Updating to the latest versions fixes
the issue.